freeze migrations into v0.21.1-beta_0.ts, bump to 0.21.1-beta:1
